+++ src/utils/variables.ts
@@ -428,5 +428,3 @@ for (const height of heights) {
 	variableOptions.push({ value: `wind_v_component_${height}m`, label: `Wind (${height}m)` });
 	variableOptions.push({ value: `wind_speed_${height}m`, label: `Wind (${height}m)` });
 }
-
-export const drawOnTiles = ['pressure_msl'];
